libpropertyinfoparser: only use system_shared_libs for bionic variant
libpropertyinfoparser is a dependency of libc, and so cannot use the
default system_shared_libs that includes libc.  It's not a dependency
of musl's libc though, and it needs to use musl's libc headers,
not the bionic libc headers provided by the libc_headers module.  Move
system_shared_libs and libc_headers into a target.bionic clause.

Bug: 190084016
Test: m USE_HOST_MUSL=true
Change-Id: Ie1f908c02fc69ae7d46b10e2fdb5dafbc3ad8557