init: update paths need to umount after first umount failure
When there are two identical mount points, the UmountPartitions()
function will recognize them as two different mount points, and will try
to umount for twice. When the first mount point is umounted, the second
mount point's umount will fail with INVALID ARGUMENT error.

The identical mount points could happen in the following situation:
"
1. mount -rbind /data/a/ /data/b/
2. mount -rbind /data/c/ /data/a/0
"
If /data has shared propagation type, the /data/a/0 and /data/b/0 are
identical mount points.

This change updates the list of partitions to unmount after the first
attempt. If the updated lists are empty, the unmount process is
considered successful. This prevents unnecessary retries and more
accurately reflects the state of the unmount operation.

Bug: 409196828
Test: Verify when /data is umounted successfully at first time, there's
no sleep & retry for umounting partitions.

Change-Id: I60dc79f392325020086e1d390357536d4037d136